Separating `FilterList` into its own directory.
